migrate (nearly) all interaction to node websockets
