+++ compositor-proxy/package.json
@@ -42,7 +42,6 @@
     "@types/js-yaml": "^4.0.5",
     "@types/jsonwebtoken": "^9.0.2",
     "@types/node": "^20.1.2",
-    "@types/supertest": "^2.0.12",
     "@types/ws": "^8.5.5",
     "@typescript-eslint/eslint-plugin": "^5.59.5",
     "@typescript-eslint/parser": "^5.59.5",
@@ -54,7 +53,6 @@
     "node-api-headers": "^1.0.0",
     "pkg": "^5.8.1",
     "prettier": "^2.8.8",
-    "supertest": "^6.3.3",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4",
